pulled component helper functions out into modules
