--- answer-log.txt
@@ -1,74 +1,0 @@
-*javascript*
-
-props
-const
-array
-https
-error
-throw
-catch
-typed
-false
-global
-local
-regex
-parse
-async
-await
-class
-comma
-super
-unary
-image
-share
-index
-video
-audio
-webgl
-keyed
-loops
-fetch
-chain
-event
-param
-brace
-scope
-arrow
-label
-empty
-break
-block
-while
-value
-
-
-*html*
-
-title
-embed
-aside
-input
-meter
-param
-small
-style
-table
-tbody
-thead
-track
-
-*input types*
-
-email
-color
-image
-month
-radio
-range
-reset
-
-*css*
-fonts
-width
-lists
-links
