+++ src/components/Game/Keyboard.js
@@ -1,4 +1,5 @@
 import React, { useState, useEffect } from "react";
+import { getAllAttemptedChars, rowScore } from '../../modules/keyboard-mgmt';
 
 import BackKey from "../UI/BackKey";
 import EnterKey from "../UI/EnterKey";
@@ -119,22 +120,6 @@ const Keyboard = (props) => {
   );
 };
 
-// =================================================================================
 
-const getAllAttemptedChars = (gState) => {
-  const attempted = gState.map((roundState, i) =>
-    roundState.map((el, j) => el[0]).filter((el) => el !== "")
-  );
-
-  const allAttemptedChars = [];
-  attempted.forEach((el) => el.forEach((char) => allAttemptedChars.push(char)));
-
-  return allAttemptedChars;
-};
-
-const rowScore = (attemptCharRow) => {
-  const nums = attemptCharRow.filter((el) => typeof el === "number");
-  return nums.reduce((prev, curr) => prev + curr);
-};
 
 export default Keyboard;
